# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee Management API",
        "description": "API para gestión de empleados y familia inmediata con sistema de filtrado avanzado",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/immediate-family": {
            "get": {
                "tags": [
                    "immediate-family"
                ],
                "summary": "Obtener todos los familiares inmediatos",
                "description": "Obtiene una lista paginada de todos los familiares inmediatos",
                "operationId": "2a8312c17e02bd8694ac0e9a08801cdf",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de familiares inmediatos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedImmediateFamilyResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en los parámetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "immediate-family"
                ],
                "summary": "Crear un nuevo familiar inmediato",
                "description": "Crea un nuevo familiar inmediato para un empleado",
                "operationId": "c9fa5c1e781d314eb2acf533f2c759e3",
                "requestBody": {
                    "description": "Datos del familiar inmediato a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id",
                                    "relative_name",
                                    "relationship",
                                    "birth_date"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "description": "ID del empleado",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "relative_name": {
                                        "description": "Nombre del familiar",
                                        "type": "string",
                                        "example": "María García"
                                    },
                                    "relationship": {
                                        "description": "Relación familiar",
                                        "type": "string",
                                        "example": "esposa"
                                    },
                                    "birth_date": {
                                        "description": "Fecha de nacimiento",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Familiar inmediato creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImmediateFamily"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/immediate-family/{id}": {
            "get": {
                "tags": [
                    "immediate-family"
                ],
                "summary": "Obtener un familiar inmediato específico",
                "description": "Obtiene los detalles de un familiar inmediato por su ID",
                "operationId": "647ecdddda812b97348ae49c583db2a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del familiar inmediato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Familiar inmediato encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImmediateFamily"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Familiar inmediato no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "immediate-family"
                ],
                "summary": "Actualizar un familiar inmediato",
                "description": "Actualiza los datos de un familiar inmediato existente",
                "operationId": "75ebdfc3fcd378e2cb38b04ceb1b1df2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del familiar inmediato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del familiar inmediato a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee_id": {
                                        "description": "ID del empleado",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "relative_name": {
                                        "description": "Nombre del familiar",
                                        "type": "string",
                                        "example": "María García"
                                    },
                                    "relationship": {
                                        "description": "Relación familiar",
                                        "type": "string",
                                        "example": "esposa"
                                    },
                                    "birth_date": {
                                        "description": "Fecha de nacimiento",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Familiar inmediato actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImmediateFamily"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Familiar inmediato no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "immediate-family"
                ],
                "summary": "Eliminar un familiar inmediato",
                "description": "Elimina un familiar inmediato del sistema",
                "operationId": "850095e03b8d78ff02a800fb94ddf445",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del familiar inmediato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Familiar inmediato eliminado exitosamente"
                    },
                    "404": {
                        "description": "Familiar inmediato no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employees/{employeeId}/immediate-family": {
            "get": {
                "tags": [
                    "immediate-family"
                ],
                "summary": "Obtener familiares inmediatos por empleado",
                "description": "Obtiene todos los familiares inmediatos de un empleado específico",
                "operationId": "28be0cc5301b0dcbfcbdb97f9f222c94",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de familiares inmediatos del empleado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ImmediateFamily"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empleado no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Employee": {
                "title": "Employee",
                "description": "Modelo de empleado",
                "required": [
                    "name",
                    "email",
                    "position",
                    "hire_date"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del empleado",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre completo del empleado",
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "description": "Correo electrónico del empleado",
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@company.com"
                    },
                    "position": {
                        "description": "Cargo o posición del empleado",
                        "type": "string",
                        "example": "Desarrollador Senior"
                    },
                    "hire_date": {
                        "description": "Fecha de contratación",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-15"
                    },
                    "created_at": {
                        "description": "Fecha de creación del registro",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-15T10:30:00Z"
                    },
                    "immediate_family": {
                        "description": "Lista de familiares inmediatos",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImmediateFamily"
                        }
                    }
                },
                "type": "object"
            },
            "ImmediateFamily": {
                "title": "ImmediateFamily",
                "description": "Modelo de familia inmediata",
                "required": [
                    "employee_id",
                    "relative_name",
                    "relationship",
                    "birth_date"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del familiar",
                        "type": "integer",
                        "example": 1
                    },
                    "employee_id": {
                        "description": "ID del empleado al que pertenece",
                        "type": "integer",
                        "example": 1
                    },
                    "relative_name": {
                        "description": "Nombre del familiar",
                        "type": "string",
                        "example": "María Pérez"
                    },
                    "relationship": {
                        "description": "Relación con el empleado",
                        "type": "string",
                        "enum": [
                            "spouse",
                            "child",
                            "parent",
                            "sibling"
                        ],
                        "example": "spouse"
                    },
                    "birth_date": {
                        "description": "Fecha de nacimiento",
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-20"
                    },
                    "created_at": {
                        "description": "Fecha de creación del registro",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-15T10:30:00Z"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    }
                },
                "type": "object"
            },
...
            "PaginatedImmediateFamilyResponse": {
                "title": "PaginatedImmediateFamilyResponse",
                "description": "Respuesta paginada de familia inmediata",
                "properties": {
                    "current_page": {
                        "description": "Página actual",
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "description": "Lista de familiares",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImmediateFamily"
                        }
                    },
                    "total": {
                        "description": "Total de elementos",
                        "type": "integer",
                        "example": 25
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Respuesta de error",
                "properties": {
                    "message": {
                        "description": "Mensaje de error",
                        "type": "string",
                        "example": "Recurso no encontrado"
                    },
                    "errors": {
                        "description": "Detalles de errores de validación",
                        "type": "object",
                        "example": {
                            "email": [
                                "El campo email es requerido"
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación JWT Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}